refactor(edge): extract listener runtime endpoint modules
