+++ crates/edge/src/quic_listener/control_api.rs
@@ -1,9 +1,13 @@
 use super::*;
+use hyper::server::conn::http1;
+use hyper::service::service_fn;
+use hyper_util::rt::TokioIo;
 use spooky_config::config::ControlApi as ControlApiConfig;
 use spooky_config::loader::read_config;
 use spooky_config::runtime::RuntimeConfig;
 use std::ffi::OsString;
 use subtle::ConstantTimeEq;
+use tokio_rustls::TlsAcceptor;
 
 #[derive(Clone)]
 pub(super) struct ControlApiPaths {
