+++ crates/edge/src/quic_listener/mod.rs
@@ -23,12 +23,6 @@ use bytes::Bytes;
 use http::{Request, Response, StatusCode};
 use http_body_util::{BodyExt, Full, combinators::BoxBody};
 use hyper::body::{Body, Frame, Incoming};
-use hyper::client::conn::http1 as client_http1;
-use hyper::server::conn::http1;
-use hyper::server::conn::http2;
-use hyper::service::service_fn;
-use hyper::upgrade;
-use hyper_util::rt::TokioIo;
 use log::{debug, error, info, warn};
 use quiche::Config;
 use quiche::h3::NameValue;
@@ -42,21 +36,18 @@ use rustls::{
 use rustls_pki_types::pem::PemObject;
 use serde_json::json;
 use socket2::{Domain, Protocol, Socket, Type};
-use spooky_bridge::h3_to_h2::{
-    ForwardedContext, ForwardedHeaderChains, build_forwarded_header_values,
-    build_h2_request_for_endpoint_with_host_policy, resolve_upstream_host_value,
-};
+use spooky_bridge::h3_to_h2::{ForwardedContext, build_h2_request_for_endpoint_with_host_policy};
 use spooky_errors::{PoolError, ProxyError, is_retryable};
 use spooky_lb::{HealthFailureReason, HealthTransition, UpstreamPool};
 use spooky_transport::h2_client::{SharedDnsResolver, TlsClientConfig};
 use spooky_transport::transport_pool::{BackendTransportKind, UpstreamTransportPool};
-use std::sync::atomic::AtomicBool;
 use tokio::runtime::Handle;
 use tokio::sync::{
     Semaphore, mpsc,
     mpsc::error::{TryRecvError, TrySendError},
     oneshot,
 };
+#[cfg(test)]
 use tokio_rustls::TlsAcceptor;
 use tracing::{Instrument, info_span};
 
@@ -93,13 +84,18 @@ use crate::{
 };
 
 mod backend_resolution;
+mod bootstrap_tls;
 mod connection;
 mod control_api;
 mod forwarding;
 mod health_check;
+mod metrics_endpoint;
+mod runtime_endpoint;
 mod token_bucket;
 mod validation;
 
+#[cfg(test)]
+use bootstrap_tls::BootstrapStartupState;
 use connection::resolve_primary_from_radix_prefix;
 pub(crate) use connection::{ConnectionRoutes, purge_connection_routes, sweep_closed_connections};
 use forwarding::{ForwardRequestMeta, abort_stream};
@@ -390,66 +386,6 @@ type BootstrapServiceFuture = std::pin::Pin<
 >;
 
 type LbHeaderLookup<'a> = dyn Fn(&str) -> Option<String> + 'a;
-
-struct BootstrapConnectionState {
-    alt_svc_value: String,
-    backend_timeout: Duration,
-    max_request_body_bytes: usize,
-    max_response_body_bytes: usize,
-    max_connections: usize,
-    connection_timeout: Duration,
-    listener_tls_store: Arc<ListenerTlsReloadStore>,
-    transport_pool: Arc<UpstreamTransportPool>,
-    backend_endpoints: Arc<HashMap<String, BackendEndpoint>>,
-    backend_resolution_store: Arc<RuntimeBackendResolutionStore>,
-    upstream_policies: Arc<HashMap<String, RuntimeUpstreamPolicy>>,
-    metrics: Arc<Metrics>,
-    resilience: Arc<RuntimeResilience>,
-    upstream_pools: HashMap<String, Arc<RwLock<UpstreamPool>>>,
-    routing_index: Arc<RouteIndex>,
-}
-
-struct BootstrapStartupState {
-    listener_config: ListenerRuntimeConfig,
-    listener_tls_store: Arc<ListenerTlsReloadStore>,
-    transport_pool: Arc<UpstreamTransportPool>,
-    backend_endpoints: Arc<HashMap<String, BackendEndpoint>>,
-    backend_resolution_store: Arc<RuntimeBackendResolutionStore>,
-    upstream_policies: Arc<HashMap<String, RuntimeUpstreamPolicy>>,
-    metrics: Arc<Metrics>,
-    resilience: Arc<RuntimeResilience>,
-    upstream_pools: HashMap<String, Arc<RwLock<UpstreamPool>>>,
-    routing_index: Arc<RouteIndex>,
-}
-
-struct MetricsEndpointState {
-    metrics_path: String,
-    max_connections: usize,
-    connection_timeout: Duration,
-    metrics: Arc<Metrics>,
-}
-
-struct MetricsEndpointBinding {
-    bind: String,
-    listener: tokio::net::TcpListener,
-    active_connections: Arc<AtomicUsize>,
-}
-
-struct RuntimeConnectionSlotGuard {
-    active_connections: Arc<AtomicUsize>,
-}
-
-impl RuntimeConnectionSlotGuard {
-    fn new(active_connections: Arc<AtomicUsize>) -> Self {
-        Self { active_connections }
-    }
-}
-
-impl Drop for RuntimeConnectionSlotGuard {
-    fn drop(&mut self) {
-        self.active_connections.fetch_sub(1, Ordering::AcqRel);
-    }
-}
 
 struct BootstrapStreamingBody {
     inner: Incoming,
@@ -4485,248 +4421,6 @@ impl QUICListener {
         default_key
     }
 
-    fn spawn_metrics_endpoint(
-        config: &RuntimeConfig,
-        metrics: Arc<Metrics>,
-        runtime_bundle: Option<Arc<RuntimeBundleHandle>>,
-    ) -> Result<(), ProxyError> {
-        let endpoint = &config.observability.metrics;
-        if runtime_bundle.is_none() && !endpoint.enabled {
-            return Ok(());
-        }
-        let required = endpoint.required;
-        let startup_endpoint = endpoint.clone();
-        let startup_metrics_path = startup_endpoint.path.clone();
-        let startup_max_connections = startup_endpoint.max_connections.max(1);
-        let startup_connection_timeout =
-            Duration::from_millis(startup_endpoint.connection_timeout_ms.max(1));
-
-        let handle = match runtime_handle() {
-            Some(handle) => handle,
-            None => {
-                let msg = "metrics endpoint disabled (no Tokio runtime available)".to_string();
-                if required {
-                    return Err(ProxyError::Transport(msg));
-                }
-                error!("{}", msg);
-                return Ok(());
-            }
-        };
-
-        let initial_binding = if startup_endpoint.enabled {
-            let bind = format!("{}:{}", startup_endpoint.address, startup_endpoint.port);
-            match Self::bind_tcp_listener(&bind, Some(&handle), "metrics endpoint") {
-                Ok(listener) => Some(MetricsEndpointBinding {
-                    bind,
-                    listener,
-                    active_connections: Arc::new(AtomicUsize::new(0)),
-                }),
-                Err(msg) => {
-                    if required {
-                        return Err(ProxyError::Transport(msg));
-                    }
-                    error!("{}", msg);
-                    None
-                }
-            }
-        } else {
-            None
-        };
-
-        spawn_supervised_async_task(
-            &handle,
-            "metrics-endpoint",
-            Some(Arc::clone(&metrics)),
-            async move {
-                let mut listener_binding = initial_binding;
-
-                loop {
-                    let endpoint = Self::current_metrics_endpoint_config(
-                        runtime_bundle.as_ref(),
-                        &startup_endpoint,
-                    );
-                    let desired_bind = format!("{}:{}", endpoint.address, endpoint.port);
-
-                    if !endpoint.enabled {
-                        if let Some(binding) = listener_binding.take() {
-                            info!(
-                                "Metrics endpoint disabled via runtime reload on {}",
-                                binding.bind
-                            );
-                        }
-                        tokio::time::sleep(Duration::from_millis(200)).await;
-                        continue;
-                    }
-
-                    let needs_rebind = match listener_binding.as_ref() {
-                        Some(binding) => binding.bind != desired_bind,
-                        None => true,
-                    };
-                    if needs_rebind {
-                        match Self::bind_tcp_listener(&desired_bind, None, "metrics endpoint") {
-                            Ok(listener) => {
-                                info!(
-                                    "Metrics endpoint listening on http://{}{} (max_connections={}, connection_timeout_ms={})",
-                                    desired_bind,
-                                    endpoint.path,
-                                    endpoint.max_connections.max(1),
-                                    endpoint.connection_timeout_ms.max(1)
-                                );
-                                listener_binding = Some(MetricsEndpointBinding {
-                                    bind: desired_bind.clone(),
-                                    listener,
-                                    active_connections: Arc::new(AtomicUsize::new(0)),
-                                });
-                            }
-                            Err(err) => {
-                                error!("{}", err);
-                                tokio::time::sleep(Duration::from_millis(200)).await;
-                                continue;
-                            }
-                        }
-                    }
-
-                    let Some(binding) = listener_binding.as_mut() else {
-                        tokio::time::sleep(Duration::from_millis(200)).await;
-                        continue;
-                    };
-
-                    let accept_result = tokio::select! {
-                        accept = binding.listener.accept() => Some(accept),
-                        _ = tokio::time::sleep(Duration::from_millis(200)) => None,
-                    };
-                    let Some(accept_result) = accept_result else {
-                        continue;
-                    };
-                    let (stream, _peer) = match accept_result {
-                        Ok(v) => v,
-                        Err(err) => {
-                            error!("Metrics endpoint accept failed: {}", err);
-                            continue;
-                        }
-                    };
-                    let runtime_state = Self::metrics_endpoint_state(
-                        runtime_bundle.as_ref(),
-                        startup_metrics_path.clone(),
-                        startup_max_connections,
-                        startup_connection_timeout,
-                        Arc::clone(&metrics),
-                    );
-                    let active_connections = Arc::clone(&binding.active_connections);
-                    if !Self::try_claim_runtime_connection_slot(
-                        &active_connections,
-                        runtime_state.max_connections,
-                    ) {
-                        continue;
-                    }
-
-                    let io = TokioIo::new(stream);
-                    let metrics = Arc::clone(&runtime_state.metrics);
-                    let metrics_path = runtime_state.metrics_path.clone();
-                    let timeout = runtime_state.connection_timeout;
-
-                    tokio::spawn(async move {
-                        let _connection_guard = RuntimeConnectionSlotGuard::new(active_connections);
-                        let service = service_fn(move |req: Request<Incoming>| {
-                            let metrics = Arc::clone(&metrics);
-                            let metrics_path = metrics_path.clone();
-                            async move {
-                                Ok::<_, hyper::Error>(Self::handle_metrics_request(
-                                    req,
-                                    &metrics_path,
-                                    metrics,
-                                ))
-                            }
-                        });
-
-                        let serve = http1::Builder::new().serve_connection(io, service);
-                        match tokio::time::timeout(timeout, serve).await {
-                            Ok(Ok(())) => {}
-                            Ok(Err(err)) => {
-                                error!("Metrics endpoint connection failed: {}", err);
-                            }
-                            Err(_) => {
-                                debug!("Metrics endpoint connection timed out");
-                            }
-                        }
-                    });
-                }
-            },
-        );
-        Ok(())
-    }
-
-    fn bind_tcp_listener(
-        bind: &str,
-        handle: Option<&Handle>,
-        context: &str,
-    ) -> Result<tokio::net::TcpListener, String> {
-        let std_listener = std::net::TcpListener::bind(bind)
-            .map_err(|err| format!("failed to bind {context} {bind}: {err}"))?;
-        std_listener.set_nonblocking(true).map_err(|err| {
-            format!(
-                "failed to set {context} listener nonblocking ({}): {}",
-                bind, err
-            )
-        })?;
-        let from_std_result = if let Some(handle) = handle {
-            let _guard = handle.enter();
-            tokio::net::TcpListener::from_std(std_listener)
-        } else {
-            tokio::net::TcpListener::from_std(std_listener)
-        };
-        from_std_result
-            .map_err(|err| format!("failed to register {context} listener {}: {}", bind, err))
-    }
-
-    fn probe_tcp_bind(bind: &str, context: &str) -> Result<(), String> {
-        std::net::TcpListener::bind(bind)
-            .map(|_| ())
-            .map_err(|err| format!("failed to bind {context} {bind}: {err}"))
-    }
-
-    fn current_metrics_endpoint_config(
-        runtime_bundle: Option<&Arc<RuntimeBundleHandle>>,
-        startup_endpoint: &spooky_config::config::MetricsEndpoint,
-    ) -> spooky_config::config::MetricsEndpoint {
-        runtime_bundle
-            .map(|handle| {
-                handle
-                    .current()
-                    .runtime_config
-                    .observability
-                    .metrics
-                    .clone()
-            })
-            .unwrap_or_else(|| startup_endpoint.clone())
-    }
-
-    fn metrics_endpoint_state(
-        runtime_bundle: Option<&Arc<RuntimeBundleHandle>>,
-        startup_metrics_path: String,
-        startup_max_connections: usize,
-        startup_connection_timeout: Duration,
-        startup_metrics: Arc<Metrics>,
-    ) -> MetricsEndpointState {
-        if let Some(handle) = runtime_bundle {
-            let runtime = handle.current();
-            let endpoint = &runtime.runtime_config.observability.metrics;
-            return MetricsEndpointState {
-                metrics_path: endpoint.path.clone(),
-                max_connections: endpoint.max_connections.max(1),
-                connection_timeout: Duration::from_millis(endpoint.connection_timeout_ms.max(1)),
-                metrics: runtime.shared_state.metrics.clone(),
-            };
-        }
-
-        MetricsEndpointState {
-            metrics_path: startup_metrics_path,
-            max_connections: startup_max_connections,
-            connection_timeout: startup_connection_timeout,
-            metrics: startup_metrics,
-        }
-    }
-
     fn load_tls_cert_chain_from_pem_file(
         path: &str,
         field_name: &str,
@@ -5267,986 +4961,6 @@ impl QUICListener {
             reason,
             reason_text
         );
-    }
-
-    pub fn spawn_bootstrap_tls_listener(
-        config: &ListenerRuntimeConfig,
-        shared_state: &SharedRuntimeState,
-        runtime_bundle: Option<Arc<RuntimeBundleHandle>>,
-        shutdown_signal: Option<Arc<AtomicBool>>,
-    ) -> Result<(), ProxyError> {
-        let bind = format!(
-            "{}:{}",
-            config.listen.listen.address, config.listen.listen.port
-        );
-        let alt_svc_value = format!("h3=\":{}\"; ma=86400", config.listen.listen.port);
-        let max_connections = config.performance.max_active_connections.max(1);
-        let connection_timeout =
-            Duration::from_millis(config.performance.client_body_idle_timeout_ms.max(1));
-        let listener_label = Self::listener_label(config);
-        shared_state
-            .listener_tls_store
-            .bootstrap_server_config(&listener_label)
-            .ok_or_else(|| {
-                ProxyError::Tls(format!(
-                    "failed to initialize bootstrap TLS listener config for '{}': missing reload state",
-                    listener_label
-                ))
-            })?;
-
-        let transport_pool = Arc::clone(&shared_state.transport_pool);
-        let backend_endpoints = Arc::clone(&shared_state.backend_endpoints);
-        let backend_resolution_store = Arc::clone(&shared_state.backend_resolution_store);
-        let upstream_policies = Arc::clone(&shared_state.upstream_policies);
-        let metrics = Arc::clone(&shared_state.metrics);
-        let resilience = Arc::clone(&shared_state.resilience);
-        let upstream_pools = shared_state.upstream_pools.clone();
-        let listener_tls_store = Arc::clone(&shared_state.listener_tls_store);
-        let runtime_bundle = runtime_bundle.clone();
-        let handle = match runtime_handle() {
-            Some(h) => h,
-            None => {
-                return Err(ProxyError::Transport(
-                    "failed to start bootstrap TLS listener: no Tokio runtime available"
-                        .to_string(),
-                ));
-            }
-        };
-
-        let std_listener = std::net::TcpListener::bind(&bind).map_err(|err| {
-            ProxyError::Transport(format!(
-                "failed to bind bootstrap TLS listener on {}: {}",
-                bind, err
-            ))
-        })?;
-        if let Err(err) = std_listener.set_nonblocking(true) {
-            return Err(ProxyError::Transport(format!(
-                "failed to set bootstrap TLS listener nonblocking ({}): {}",
-                bind, err
-            )));
-        }
-        let listener = {
-            let _guard = handle.enter();
-            tokio::net::TcpListener::from_std(std_listener).map_err(|err| {
-                ProxyError::Transport(format!(
-                    "failed to register bootstrap TLS listener {}: {}",
-                    bind, err
-                ))
-            })?
-        };
-
-        let startup_state = BootstrapStartupState {
-            listener_config: config.clone(),
-            listener_tls_store: Arc::clone(&listener_tls_store),
-            transport_pool: Arc::clone(&transport_pool),
-            backend_endpoints: Arc::clone(&backend_endpoints),
-            backend_resolution_store: Arc::clone(&backend_resolution_store),
-            upstream_policies: Arc::clone(&upstream_policies),
-            metrics: Arc::clone(&metrics),
-            resilience: Arc::clone(&resilience),
-            upstream_pools: upstream_pools.clone(),
-            routing_index: Arc::clone(&shared_state.routing_index),
-        };
-
-        spawn_supervised_async_task(&handle, "bootstrap-tls-listener", None, async move {
-            info!(
-                "Bootstrap TLS listener on https://{} (TCP+TLS) — advertising Alt-Svc: {} (max_connections={}, connection_timeout_ms={})",
-                bind,
-                alt_svc_value,
-                max_connections,
-                connection_timeout.as_millis()
-            );
-            let active_connections = Arc::new(AtomicUsize::new(0));
-            loop {
-                let accept_result = if let Some(shutdown_signal) = shutdown_signal.as_ref() {
-                    tokio::select! {
-                        accept = listener.accept() => Some(accept),
-                        _ = tokio::time::sleep(Duration::from_millis(200)) => {
-                            if shutdown_signal.load(Ordering::Relaxed) {
-                                None
-                            } else {
-                                continue;
-                            }
-                        }
-                    }
-                } else {
-                    Some(listener.accept().await)
-                };
-                let Some(accept_result) = accept_result else {
-                    info!(
-                        "Bootstrap TLS listener on {} stopping due to runtime group shutdown",
-                        bind
-                    );
-                    break;
-                };
-                let (stream, peer) = match accept_result {
-                    Ok(v) => v,
-                    Err(err) => {
-                        error!("Bootstrap TLS listener accept failed: {}", err);
-                        continue;
-                    }
-                };
-                let Some(runtime_state) = Self::bootstrap_connection_state(
-                    &listener_label,
-                    runtime_bundle.as_ref(),
-                    &startup_state,
-                ) else {
-                    error!(
-                        "Bootstrap TLS listener missing live runtime state for listener {}",
-                        listener_label
-                    );
-                    continue;
-                };
-                let active_connections = Arc::clone(&active_connections);
-                if !Self::try_claim_runtime_connection_slot(
-                    &active_connections,
-                    runtime_state.max_connections,
-                ) {
-                    runtime_state.metrics.inc_connection_cap_reject();
-                    debug!(
-                        "Bootstrap TLS listener dropped connection from {}: max_connections reached",
-                        peer
-                    );
-                    continue;
-                }
-
-                let alt_svc = runtime_state.alt_svc_value.clone();
-                let transport_pool = Arc::clone(&runtime_state.transport_pool);
-                let backend_endpoints = Arc::clone(&runtime_state.backend_endpoints);
-                let backend_resolution_store = Arc::clone(&runtime_state.backend_resolution_store);
-                let upstream_policies = Arc::clone(&runtime_state.upstream_policies);
-                let metrics = Arc::clone(&runtime_state.metrics);
-                let resilience = Arc::clone(&runtime_state.resilience);
-                let upstream_pools = runtime_state.upstream_pools.clone();
-                let routing_index = Arc::clone(&runtime_state.routing_index);
-                let max_request_body_bytes = runtime_state.max_request_body_bytes;
-                let max_response_body_bytes = runtime_state.max_response_body_bytes;
-                let backend_timeout = runtime_state.backend_timeout;
-                let timeout = runtime_state.connection_timeout;
-                let listener_label = listener_label.clone();
-                let listener_tls_store = Arc::clone(&runtime_state.listener_tls_store);
-
-                tokio::spawn(async move {
-                    let _connection_guard = RuntimeConnectionSlotGuard::new(active_connections);
-                    let Some(server_config) =
-                        listener_tls_store.bootstrap_server_config(&listener_label)
-                    else {
-                        error!(
-                            "Bootstrap TLS listener missing live server config for listener {}",
-                            listener_label
-                        );
-                        return;
-                    };
-                    let acceptor = TlsAcceptor::from(server_config);
-                    let tls_stream = match acceptor.accept(stream).await {
-                        Ok(s) => s,
-                        Err(err) => {
-                            let err_text = err.to_string();
-                            let reason = Self::classify_downstream_tls_failure_reason(&err_text);
-                            metrics
-                                .record_downstream_tls_handshake_failure(&listener_label, reason);
-                            debug!(
-                                "Bootstrap TLS handshake failed listener={} peer={} reason={} error={}",
-                                listener_label, peer, reason, err_text
-                            );
-                            return;
-                        }
-                    };
-
-                    let Some(listener_tls) = listener_tls_store.inventory(&listener_label) else {
-                        error!(
-                            "Bootstrap TLS listener missing live inventory for listener {}",
-                            listener_label
-                        );
-                        return;
-                    };
-                    let requested_sni = tls_stream.get_ref().1.server_name().map(str::to_string);
-                    let (selection, identity) = Self::classify_downstream_tls_cert_selection(
-                        &listener_tls.listener_tls,
-                        requested_sni.as_deref(),
-                    );
-                    let negotiated = tls_stream.get_ref().1.alpn_protocol().map(|p| p.to_vec());
-                    let negotiated_label = negotiated
-                        .as_deref()
-                        .and_then(|value| std::str::from_utf8(value).ok())
-                        .unwrap_or("none");
-                    let client_cert_present = tls_stream
-                        .get_ref()
-                        .1
-                        .peer_certificates()
-                        .is_some_and(|certs| !certs.is_empty());
-                    metrics.inc_downstream_tls_handshake_success();
-                    metrics.record_downstream_tls_cert_selection(&listener_label, selection);
-                    metrics.record_downstream_tls_alpn(&listener_label, negotiated_label);
-                    debug!(
-                        "Bootstrap TLS handshake success listener={} peer={} sni={:?} selection={} cert='{}' alpn={} client_cert_present={}",
-                        listener_label,
-                        peer,
-                        requested_sni,
-                        selection,
-                        identity.cert_path,
-                        negotiated_label,
-                        client_cert_present
-                    );
-                    let use_h2 = negotiated.as_deref() == Some(b"h2");
-
-                    let io = TokioIo::new(tls_stream);
-                    let alt_svc_conn = alt_svc.clone();
-
-                    let svc = service_fn(
-                        move |mut req: Request<Incoming>| -> BootstrapServiceFuture {
-                            let alt = alt_svc_conn.clone();
-                            let transport_pool = Arc::clone(&transport_pool);
-                            let backend_endpoints = Arc::clone(&backend_endpoints);
-                            let backend_resolution_store = Arc::clone(&backend_resolution_store);
-                            let upstream_policies = Arc::clone(&upstream_policies);
-                            let metrics = Arc::clone(&metrics);
-                            let resilience = Arc::clone(&resilience);
-                            let upstream_pools = upstream_pools.clone();
-                            let routing_index = Arc::clone(&routing_index);
-                            let max_request_body_bytes = max_request_body_bytes;
-                            let max_response_body_bytes = max_response_body_bytes;
-
-                            Box::pin(async move {
-                                let is_websocket_upgrade =
-                                    is_websocket_upgrade_request(&req, use_h2);
-                                let client_upgrade = if is_websocket_upgrade {
-                                    Some(upgrade::on(&mut req))
-                                } else {
-                                    None
-                                };
-
-                                let request = match validate_http_request(&req, &resilience) {
-                                    Ok(request) => request,
-                                    Err((status, body, is_policy)) => {
-                                        metrics.inc_failure();
-                                        metrics.inc_request_validation_reject();
-                                        if is_policy {
-                                            metrics.inc_policy_denied();
-                                        }
-                                        metrics.record_route(
-                                            "unrouted",
-                                            Duration::from_millis(0),
-                                            RouteOutcome::Failure,
-                                        );
-                                        return Ok(Response::builder()
-                                            .status(status)
-                                            .header("alt-svc", &alt)
-                                            .body(boxed_full(Bytes::copy_from_slice(body)))
-                                            .unwrap_or_else(|_| {
-                                                Response::new(boxed_full(Bytes::new()))
-                                            }));
-                                    }
-                                };
-                                let method = request.method;
-                                let path = request.path;
-                                let authority = request.authority;
-                                let content_length = request.content_length;
-                                let suppress_downstream_body = is_head_method(&method);
-
-                                let bootstrap_error = |status: StatusCode, body: &'static [u8]| {
-                                    Ok(Response::builder()
-                                        .status(status)
-                                        .header("alt-svc", &alt)
-                                        .body(boxed_full(Bytes::from_static(body)))
-                                        .unwrap_or_else(|_| {
-                                            Response::new(boxed_full(Bytes::from_static(
-                                                b"error\n",
-                                            )))
-                                        }))
-                                };
-
-                                // Preserve the same route-lookup + tie-break semantics as the QUIC
-                                // data plane by delegating bootstrap backend resolution here.
-                                let lb_header_lookup = |name: &str| {
-                                    req.headers()
-                                        .get(name)
-                                        .and_then(|value| value.to_str().ok())
-                                        .map(str::to_string)
-                                };
-                                let resolved = Self::resolve_backend(
-                                    &method,
-                                    &path,
-                                    authority.as_deref(),
-                                    None,
-                                    &upstream_pools,
-                                    &routing_index,
-                                    Some(&lb_header_lookup),
-                                );
-                                let (backend_addr, upstream_name) = match resolved {
-                                    Ok(value) => (value.backend_addr, value.upstream_name),
-                                    Err(ProxyError::Transport(reason)) => {
-                                        let (status, body) =
-                                            bootstrap_resolution_error_response(&reason);
-                                        if status == StatusCode::BAD_GATEWAY
-                                            && body == b"route/backend resolution failed\n"
-                                        {
-                                            warn!(
-                                                "Bootstrap route/backend resolution failed: {}",
-                                                reason
-                                            );
-                                        }
-                                        return bootstrap_error(status, body);
-                                    }
-                                    Err(err) => {
-                                        warn!("Bootstrap route/backend resolution failed: {}", err);
-                                        return bootstrap_error(
-                                            StatusCode::BAD_GATEWAY,
-                                            b"route/backend resolution failed\n",
-                                        );
-                                    }
-                                };
-
-                                let endpoint = match backend_endpoints.get(&backend_addr) {
-                                    Some(ep) => ep.clone(),
-                                    None => {
-                                        return Ok(Response::builder()
-                                            .status(StatusCode::BAD_GATEWAY)
-                                            .header("alt-svc", &alt)
-                                            .body(boxed_full(Bytes::from_static(b"no endpoint\n")))
-                                            .unwrap_or_else(|_| {
-                                                Response::new(boxed_full(Bytes::from_static(
-                                                    b"error\n",
-                                                )))
-                                            }));
-                                    }
-                                };
-
-                                // Build upstream request
-                                let request_path = if path.is_empty() { "/" } else { &path };
-                                let upstream_policy = upstream_policies
-                                    .get(&upstream_name)
-                                    .cloned()
-                                    .unwrap_or_default();
-                                if !is_websocket_upgrade
-                                    && content_length
-                                        .is_some_and(|value| value > max_request_body_bytes)
-                                {
-                                    return Ok(Response::builder()
-                                        .status(StatusCode::PAYLOAD_TOO_LARGE)
-                                        .header("alt-svc", &alt)
-                                        .body(boxed_full(Bytes::from_static(
-                                            b"request body too large\n",
-                                        )))
-                                        .unwrap_or_else(|_| {
-                                            Response::new(boxed_full(Bytes::from_static(
-                                                b"error\n",
-                                            )))
-                                        }));
-                                }
-
-                                let request_id = REQUEST_ID_COUNTER.fetch_add(1, Ordering::Relaxed);
-                                let traceparent = req
-                                    .headers()
-                                    .get("traceparent")
-                                    .and_then(|value| value.to_str().ok())
-                                    .map(str::to_string);
-                                let upstream_req = if is_websocket_upgrade {
-                                    let upstream_uri = match http::Uri::try_from(
-                                        endpoint.uri_for_path(request_path),
-                                    ) {
-                                        Ok(uri) => uri,
-                                        Err(_) => {
-                                            return Ok(Response::builder()
-                                                .status(StatusCode::BAD_GATEWAY)
-                                                .header("alt-svc", &alt)
-                                                .body(boxed_full(Bytes::from_static(b"bad uri\n")))
-                                                .unwrap_or_else(|_| {
-                                                    Response::new(boxed_full(Bytes::from_static(
-                                                        b"error\n",
-                                                    )))
-                                                }));
-                                        }
-                                    };
-                                    let request_host = req
-                                        .headers()
-                                        .get(http::header::HOST)
-                                        .and_then(|value| value.to_str().ok());
-                                    let upstream_host = match resolve_upstream_host_value(
-                                        &endpoint,
-                                        &upstream_policy.host.0,
-                                        authority.as_deref(),
-                                        request_host,
-                                    ) {
-                                        Ok(host) => host,
-                                        Err(_) => {
-                                            return Ok(Response::builder()
-                                                .status(StatusCode::BAD_REQUEST)
-                                                .header("alt-svc", &alt)
-                                                .body(boxed_full(Bytes::from_static(
-                                                    b"invalid host policy\n",
-                                                )))
-                                                .unwrap_or_else(|_| {
-                                                    Response::new(boxed_full(Bytes::from_static(
-                                                        b"error\n",
-                                                    )))
-                                                }));
-                                        }
-                                    };
-
-                                    let mut upstream_req = Request::builder()
-                                        .method(method.as_str())
-                                        .uri(upstream_uri);
-                                    let mut forwarded_from_headers: Vec<Vec<u8>> = Vec::new();
-                                    let mut x_forwarded_for_from_headers: Vec<Vec<u8>> = Vec::new();
-                                    let mut x_forwarded_proto_from_headers: Vec<Vec<u8>> =
-                                        Vec::new();
-                                    let mut x_forwarded_host_from_headers: Vec<Vec<u8>> =
-                                        Vec::new();
-                                    for (name, value) in req.headers() {
-                                        if name == http::header::HOST {
-                                            continue;
-                                        }
-                                        if name.as_str().eq_ignore_ascii_case("forwarded") {
-                                            forwarded_from_headers.push(value.as_bytes().to_vec());
-                                            continue;
-                                        }
-                                        if name.as_str().eq_ignore_ascii_case("x-forwarded-for") {
-                                            x_forwarded_for_from_headers
-                                                .push(value.as_bytes().to_vec());
-                                            continue;
-                                        }
-                                        if name.as_str().eq_ignore_ascii_case("x-forwarded-proto") {
-                                            x_forwarded_proto_from_headers
-                                                .push(value.as_bytes().to_vec());
-                                            continue;
-                                        }
-                                        if name.as_str().eq_ignore_ascii_case("x-forwarded-host") {
-                                            x_forwarded_host_from_headers
-                                                .push(value.as_bytes().to_vec());
-                                            continue;
-                                        }
-                                        if name == http::header::PROXY_AUTHORIZATION
-                                            || name == http::header::PROXY_AUTHENTICATE
-                                            || name
-                                                .as_str()
-                                                .eq_ignore_ascii_case("proxy-connection")
-                                            || name == http::header::CONTENT_LENGTH
-                                            || name == http::header::TE
-                                            || name == http::header::TRAILER
-                                            || name == http::header::TRANSFER_ENCODING
-                                            || name.as_str().eq_ignore_ascii_case("keep-alive")
-                                            || name.as_str().eq_ignore_ascii_case("forwarded")
-                                            || name.as_str().eq_ignore_ascii_case("x-forwarded-for")
-                                            || name
-                                                .as_str()
-                                                .eq_ignore_ascii_case("x-forwarded-proto")
-                                            || name
-                                                .as_str()
-                                                .eq_ignore_ascii_case("x-forwarded-host")
-                                        {
-                                            continue;
-                                        }
-                                        upstream_req = upstream_req.header(name, value);
-                                    }
-                                    upstream_req =
-                                        upstream_req.header(http::header::HOST, upstream_host);
-
-                                    let forwarded_values = match build_forwarded_header_values(
-                                        &upstream_policy.forwarded_headers.0,
-                                        ForwardedHeaderChains {
-                                            forwarded: &forwarded_from_headers,
-                                            x_forwarded_for: &x_forwarded_for_from_headers,
-                                            x_forwarded_proto: &x_forwarded_proto_from_headers,
-                                            x_forwarded_host: &x_forwarded_host_from_headers,
-                                        },
-                                        peer.ip(),
-                                        upstream_host,
-                                    ) {
-                                        Ok(values) => values,
-                                        Err(err) => {
-                                            warn!(
-                                                "Bootstrap forwarded header policy failed: {}",
-                                                err
-                                            );
-                                            return Ok(Response::builder()
-                                                .status(StatusCode::BAD_REQUEST)
-                                                .header("alt-svc", &alt)
-                                                .body(boxed_full(Bytes::from_static(
-                                                    b"invalid forwarded headers\n",
-                                                )))
-                                                .unwrap_or_else(|_| {
-                                                    Response::new(boxed_full(Bytes::from_static(
-                                                        b"error\n",
-                                                    )))
-                                                }));
-                                        }
-                                    };
-                                    if let Some(value) = forwarded_values.forwarded {
-                                        upstream_req = upstream_req.header("forwarded", value);
-                                    }
-                                    if let Some(value) = forwarded_values.x_forwarded_for {
-                                        upstream_req =
-                                            upstream_req.header("x-forwarded-for", value);
-                                    }
-                                    if let Some(value) = forwarded_values.x_forwarded_proto {
-                                        upstream_req =
-                                            upstream_req.header("x-forwarded-proto", value);
-                                    }
-                                    if let Some(value) = forwarded_values.x_forwarded_host {
-                                        upstream_req =
-                                            upstream_req.header("x-forwarded-host", value);
-                                    }
-
-                                    match upstream_req.body(boxed_full(Bytes::new())) {
-                                        Ok(request) => request,
-                                        Err(_) => {
-                                            return Ok(Response::builder()
-                                                .status(StatusCode::BAD_GATEWAY)
-                                                .header("alt-svc", &alt)
-                                                .body(boxed_full(Bytes::from_static(
-                                                    b"request build error\n",
-                                                )))
-                                                .unwrap_or_else(|_| {
-                                                    Response::new(boxed_full(Bytes::from_static(
-                                                        b"error\n",
-                                                    )))
-                                                }));
-                                        }
-                                    }
-                                } else {
-                                    let bridge_headers: Vec<quiche::h3::Header> = req
-                                        .headers()
-                                        .iter()
-                                        .map(|(name, value)| {
-                                            quiche::h3::Header::new(
-                                                name.as_str().as_bytes(),
-                                                value.as_bytes(),
-                                            )
-                                        })
-                                        .collect();
-                                    match build_h2_request_for_endpoint_with_host_policy(
-                                        &endpoint,
-                                        &upstream_policy.host.0,
-                                        &upstream_policy.forwarded_headers.0,
-                                        &method,
-                                        &path,
-                                        &bridge_headers,
-                                        BootstrapStreamingBody::new(req.into_body())
-                                            .map_err(|never| match never {})
-                                            .boxed(),
-                                        None,
-                                        ForwardedContext {
-                                            client_addr: peer,
-                                            request_authority: authority.as_deref(),
-                                            request_id,
-                                            traceparent: traceparent.as_deref(),
-                                        },
-                                    ) {
-                                        Ok(request) => request,
-                                        Err(err) => {
-                                            warn!("Bootstrap request build failed: {}", err);
-                                            return Ok(Response::builder()
-                                                .status(StatusCode::BAD_REQUEST)
-                                                .header("alt-svc", &alt)
-                                                .body(boxed_full(Bytes::from_static(
-                                                    b"invalid request\n",
-                                                )))
-                                                .unwrap_or_else(|_| {
-                                                    Response::new(boxed_full(Bytes::from_static(
-                                                        b"error\n",
-                                                    )))
-                                                }));
-                                        }
-                                    }
-                                };
-                                let mut upstream_resp = if is_websocket_upgrade {
-                                    if endpoint.scheme() != BackendScheme::Http {
-                                        return Ok(Response::builder()
-                                            .status(StatusCode::BAD_GATEWAY)
-                                            .header("alt-svc", &alt)
-                                            .body(boxed_full(Bytes::from_static(
-                                                b"websocket bootstrap requires http upstream\n",
-                                            )))
-                                            .unwrap_or_else(|_| {
-                                                Response::new(boxed_full(Bytes::from_static(
-                                                    b"error\n",
-                                                )))
-                                            }));
-                                    }
-                                    let backend_target = endpoint.authority().to_string();
-                                    let upstream_path_uri = match http::Uri::try_from(request_path)
-                                    {
-                                        Ok(uri) => uri,
-                                        Err(_) => {
-                                            return Ok(Response::builder()
-                                                .status(StatusCode::BAD_GATEWAY)
-                                                .header("alt-svc", &alt)
-                                                .body(boxed_full(Bytes::from_static(b"bad uri\n")))
-                                                .unwrap_or_else(|_| {
-                                                    Response::new(boxed_full(Bytes::from_static(
-                                                        b"error\n",
-                                                    )))
-                                                }));
-                                        }
-                                    };
-                                    let (mut parts, body) = upstream_req.into_parts();
-                                    parts.uri = upstream_path_uri;
-                                    let upstream_req = Request::from_parts(parts, body);
-
-                                    let stream = match tokio::time::timeout(
-                                        backend_timeout,
-                                        tokio::net::TcpStream::connect(&backend_target),
-                                    )
-                                    .await
-                                    {
-                                        Ok(Ok(s)) => s,
-                                        Ok(Err(err)) => {
-                                            warn!("Bootstrap WebSocket connect error: {}", err);
-                                            return Ok(Response::builder()
-                                                .status(StatusCode::BAD_GATEWAY)
-                                                .header("alt-svc", &alt)
-                                                .body(boxed_full(Bytes::from_static(
-                                                    b"upstream error\n",
-                                                )))
-                                                .unwrap_or_else(|_| {
-                                                    Response::new(boxed_full(Bytes::from_static(
-                                                        b"error\n",
-                                                    )))
-                                                }));
-                                        }
-                                        Err(_) => {
-                                            return Ok(Response::builder()
-                                                .status(StatusCode::GATEWAY_TIMEOUT)
-                                                .header("alt-svc", &alt)
-                                                .body(boxed_full(Bytes::from_static(
-                                                    b"upstream timeout\n",
-                                                )))
-                                                .unwrap_or_else(|_| {
-                                                    Response::new(boxed_full(Bytes::from_static(
-                                                        b"error\n",
-                                                    )))
-                                                }));
-                                        }
-                                    };
-                                    let io = TokioIo::new(stream);
-                                    let (mut sender, conn) = match client_http1::handshake(io).await
-                                    {
-                                        Ok(v) => v,
-                                        Err(err) => {
-                                            warn!(
-                                                "Bootstrap WebSocket handshake setup failed: {}",
-                                                err
-                                            );
-                                            return Ok(Response::builder()
-                                                .status(StatusCode::BAD_GATEWAY)
-                                                .header("alt-svc", &alt)
-                                                .body(boxed_full(Bytes::from_static(
-                                                    b"upstream error\n",
-                                                )))
-                                                .unwrap_or_else(|_| {
-                                                    Response::new(boxed_full(Bytes::from_static(
-                                                        b"error\n",
-                                                    )))
-                                                }));
-                                        }
-                                    };
-                                    tokio::spawn(async move {
-                                        let _ = conn.with_upgrades().await;
-                                    });
-                                    match tokio::time::timeout(
-                                        backend_timeout,
-                                        sender.send_request(upstream_req),
-                                    )
-                                    .await
-                                    {
-                                        Ok(Ok(resp)) => resp,
-                                        Ok(Err(err)) => {
-                                            warn!("Bootstrap WebSocket upstream error: {}", err);
-                                            return Ok(Response::builder()
-                                                .status(StatusCode::BAD_GATEWAY)
-                                                .header("alt-svc", &alt)
-                                                .body(boxed_full(Bytes::from_static(
-                                                    b"upstream error\n",
-                                                )))
-                                                .unwrap_or_else(|_| {
-                                                    Response::new(boxed_full(Bytes::from_static(
-                                                        b"error\n",
-                                                    )))
-                                                }));
-                                        }
-                                        Err(_) => {
-                                            return Ok(Response::builder()
-                                                .status(StatusCode::GATEWAY_TIMEOUT)
-                                                .header("alt-svc", &alt)
-                                                .body(boxed_full(Bytes::from_static(
-                                                    b"upstream timeout\n",
-                                                )))
-                                                .unwrap_or_else(|_| {
-                                                    Response::new(boxed_full(Bytes::from_static(
-                                                        b"error\n",
-                                                    )))
-                                                }));
-                                        }
-                                    }
-                                } else {
-                                    Self::record_backend_connect_attempt(
-                                        &metrics,
-                                        &backend_resolution_store,
-                                        &backend_addr,
-                                    );
-                                    match tokio::time::timeout(
-                                        backend_timeout,
-                                        transport_pool.send(&backend_addr, upstream_req),
-                                    )
-                                    .await
-                                    {
-                                        Ok(Ok(resp)) => resp,
-                                        Ok(Err(err)) => {
-                                            warn!("Bootstrap proxy upstream error: {}", err);
-                                            return Ok(Response::builder()
-                                                .status(StatusCode::BAD_GATEWAY)
-                                                .header("alt-svc", &alt)
-                                                .body(boxed_full(Bytes::from_static(
-                                                    b"upstream error\n",
-                                                )))
-                                                .unwrap_or_else(|_| {
-                                                    Response::new(boxed_full(Bytes::from_static(
-                                                        b"error\n",
-                                                    )))
-                                                }));
-                                        }
-                                        Err(_) => {
-                                            return Ok(Response::builder()
-                                                .status(StatusCode::GATEWAY_TIMEOUT)
-                                                .header("alt-svc", &alt)
-                                                .body(boxed_full(Bytes::from_static(
-                                                    b"upstream timeout\n",
-                                                )))
-                                                .unwrap_or_else(|_| {
-                                                    Response::new(boxed_full(Bytes::from_static(
-                                                        b"error\n",
-                                                    )))
-                                                }));
-                                        }
-                                    }
-                                };
-
-                                // Build downstream response with Alt-Svc injected
-                                if !suppress_downstream_body
-                                    && let Some(content_length) = upstream_resp
-                                        .headers()
-                                        .get(http::header::CONTENT_LENGTH)
-                                        .and_then(|v| v.to_str().ok())
-                                        .and_then(|s| s.parse::<usize>().ok())
-                                    && content_length > max_response_body_bytes
-                                {
-                                    return Ok(Response::builder()
-                                        .status(StatusCode::SERVICE_UNAVAILABLE)
-                                        .header("alt-svc", &alt)
-                                        .body(boxed_full(Bytes::from_static(
-                                            b"upstream response body too large\n",
-                                        )))
-                                        .unwrap_or_else(|_| {
-                                            Response::new(boxed_full(Bytes::from_static(
-                                                b"error\n",
-                                            )))
-                                        }));
-                                }
-
-                                let status = upstream_resp.status();
-                                let mut resp_builder = Response::builder().status(status);
-                                let response_connection_tokens =
-                                    connection_header_tokens(upstream_resp.headers());
-                                for (name, value) in upstream_resp.headers() {
-                                    if should_strip_bootstrap_response_header(
-                                        name,
-                                        &response_connection_tokens,
-                                    ) {
-                                        continue;
-                                    }
-                                    resp_builder = resp_builder.header(name, value);
-                                }
-                                resp_builder = resp_builder.header("alt-svc", &alt);
-                                if is_websocket_upgrade
-                                    && upstream_resp.status() == StatusCode::SWITCHING_PROTOCOLS
-                                {
-                                    let client_upgrade = match client_upgrade {
-                                        Some(u) => u,
-                                        None => {
-                                            return Ok(Response::builder()
-                                                .status(StatusCode::BAD_GATEWAY)
-                                                .header("alt-svc", &alt)
-                                                .body(boxed_full(Bytes::from_static(
-                                                    b"upgrade setup error\n",
-                                                )))
-                                                .unwrap_or_else(|_| {
-                                                    Response::new(boxed_full(Bytes::from_static(
-                                                        b"error\n",
-                                                    )))
-                                                }));
-                                        }
-                                    };
-                                    let upstream_upgrade = upgrade::on(&mut upstream_resp);
-                                    tokio::spawn(async move {
-                                        let (client, upstream) = match tokio::try_join!(
-                                            client_upgrade,
-                                            upstream_upgrade
-                                        ) {
-                                            Ok(v) => v,
-                                            Err(err) => {
-                                                debug!(
-                                                    "Bootstrap WebSocket upgrade join failed: {}",
-                                                    err
-                                                );
-                                                return;
-                                            }
-                                        };
-                                        let mut client = TokioIo::new(client);
-                                        let mut upstream = TokioIo::new(upstream);
-                                        let _ = tokio::io::copy_bidirectional(
-                                            &mut client,
-                                            &mut upstream,
-                                        )
-                                        .await;
-                                    });
-                                    return Ok(resp_builder
-                                        .body(boxed_full(Bytes::new()))
-                                        .unwrap_or_else(|_| {
-                                            Response::new(boxed_full(Bytes::new()))
-                                        }));
-                                }
-                                let resp_body = if suppress_downstream_body {
-                                    boxed_full(Bytes::new())
-                                } else {
-                                    BootstrapStreamingBody::with_max_bytes(
-                                        upstream_resp.into_body(),
-                                        max_response_body_bytes,
-                                    )
-                                    .map_err(|never| match never {})
-                                    .boxed()
-                                };
-
-                                Ok(resp_builder
-                                    .body(resp_body)
-                                    .unwrap_or_else(|_| Response::new(boxed_full(Bytes::new()))))
-                            })
-                        },
-                    );
-
-                    if use_h2 {
-                        let executor = hyper_util::rt::TokioExecutor::new();
-                        let serve = http2::Builder::new(executor).serve_connection(io, svc);
-                        match tokio::time::timeout(timeout, serve).await {
-                            Ok(Ok(())) => {}
-                            Ok(Err(err)) => {
-                                debug!("Bootstrap h2 connection from {} closed: {}", peer, err);
-                            }
-                            Err(_) => {
-                                debug!("Bootstrap h2 connection from {} timed out", peer);
-                            }
-                        }
-                    } else {
-                        let serve = http1::Builder::new()
-                            .serve_connection(io, svc)
-                            .with_upgrades();
-                        match tokio::time::timeout(timeout, serve).await {
-                            Ok(Ok(())) => {}
-                            Ok(Err(err)) => {
-                                debug!("Bootstrap h1 connection from {} closed: {}", peer, err);
-                            }
-                            Err(_) => {
-                                debug!("Bootstrap h1 connection from {} timed out", peer);
-                            }
-                        }
-                    }
-                });
-            }
-        });
-
-        Ok(())
-    }
-
-    fn bootstrap_connection_state(
-        listener_label: &str,
-        runtime_bundle: Option<&Arc<RuntimeBundleHandle>>,
-        startup: &BootstrapStartupState,
-    ) -> Option<BootstrapConnectionState> {
-        let (
-            listener_config,
-            listener_tls_store,
-            transport_pool,
-            backend_endpoints,
-            backend_resolution_store,
-            upstream_policies,
-            metrics,
-            resilience,
-            upstream_pools,
-            routing_index,
-        ) = if let Some(handle) = runtime_bundle {
-            let runtime = handle.current();
-            (
-                runtime.listener_runtime_config(listener_label)?,
-                runtime.shared_state.listener_tls_store.clone(),
-                runtime.shared_state.transport_pool.clone(),
-                runtime.shared_state.backend_endpoints.clone(),
-                runtime.shared_state.backend_resolution_store.clone(),
-                runtime.shared_state.upstream_policies.clone(),
-                runtime.shared_state.metrics.clone(),
-                runtime.shared_state.resilience.clone(),
-                runtime.shared_state.upstream_pools.clone(),
-                runtime.shared_state.routing_index.clone(),
-            )
-        } else {
-            (
-                startup.listener_config.clone(),
-                Arc::clone(&startup.listener_tls_store),
-                Arc::clone(&startup.transport_pool),
-                Arc::clone(&startup.backend_endpoints),
-                Arc::clone(&startup.backend_resolution_store),
-                Arc::clone(&startup.upstream_policies),
-                Arc::clone(&startup.metrics),
-                Arc::clone(&startup.resilience),
-                startup.upstream_pools.clone(),
-                Arc::clone(&startup.routing_index),
-            )
-        };
-
-        Some(BootstrapConnectionState {
-            alt_svc_value: format!("h3=\":{}\"; ma=86400", listener_config.listen.listen.port),
-            backend_timeout: Duration::from_millis(listener_config.performance.backend_timeout_ms),
-            max_request_body_bytes: listener_config.performance.max_request_body_bytes,
-            max_response_body_bytes: listener_config.performance.max_response_body_bytes,
-            max_connections: listener_config.performance.max_active_connections.max(1),
-            connection_timeout: Duration::from_millis(
-                listener_config
-                    .performance
-                    .client_body_idle_timeout_ms
-                    .max(1),
-            ),
-            listener_tls_store,
-            transport_pool,
-            backend_endpoints,
-            backend_resolution_store,
-            upstream_policies,
-            metrics,
-            resilience,
-            upstream_pools,
-            routing_index,
-        })
-    }
-
-    fn try_claim_runtime_connection_slot(
-        active_connections: &Arc<AtomicUsize>,
-        max_connections: usize,
-    ) -> bool {
-        loop {
-            let current = active_connections.load(Ordering::Relaxed);
-            if current >= max_connections {
-                return false;
-            }
-            if active_connections
-                .compare_exchange(
-                    current,
-                    current.saturating_add(1),
-                    Ordering::AcqRel,
-                    Ordering::Relaxed,
-                )
-                .is_ok()
-            {
-                return true;
-            }
-        }
     }
 
     fn handle_metrics_request(
